WIP fix for syncing to autohub ('versions_url')
This is mostly result of API differences between
galaxy collections v2 and autohub v3 api.

In this case, it's that 'version_url' in
v2 is a "full" url, while in v3 it is just
the path with no scheme.

Also lots of debug logging...

Fixes https://pulp.plan.io/issues/6116